Fix silent SARIMAX failures, cumsum precision, and input validation
- Add warnings when SARIMAX predict returns 0.0 on unfitted model
- Track fit/failure counts in _SARIMAXEstimator
- Use float64 dtype for cumsum in HARFeatures to prevent precision loss
- Move circuit breaker dates to config.py for maintainability
- Change RollingBuffer dtype from float32 to float64 for consistency
- Add check_backtest_inputs helper and use in run_backtest_agnostic

+++ src/features.py
@@ -79,7 +79,7 @@ class HARFeatures(LagFeatureBase):
 
     def _compute_lag_np(self, col_array, lag):
         n = len(col_array)
-        cumsum = np.cumsum(col_array)
+        cumsum = np.cumsum(col_array, dtype=np.float64)
         rolling_mean = np.empty(n)
         for i in range(n):
             window_start = max(0, i - lag + 1)
